Add unmanaged support for keyvault and cosmos

+++ pkg/curp/handlers/azure_cosmosdb_mongo.go
@@ -13,6 +13,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/to"
 	"github.com/Azure/radius/pkg/curp/armauth"
 	"github.com/Azure/radius/pkg/curp/resources"
+	radresources "github.com/Azure/radius/pkg/curp/resources"
 	"github.com/Azure/radius/pkg/rad/util"
 )
 
@@ -31,34 +32,59 @@ type azureCosmosDBMongoHandler struct {
 func (handler *azureCosmosDBMongoHandler) Put(ctx context.Context, options PutOptions) (map[string]string, error) {
 	properties := mergeProperties(options.Resource, options.Existing)
 
-	account, err := handler.CreateCosmosDBAccount(ctx, properties, documentdb.MongoDB)
-	if err != nil {
-		return nil, err
+	// This assertion is important so we don't start creating/modifying an unmanaged resource
+	if properties[ManagedKey] != "true" && (properties[CosmosDBAccountIDKey] == "" || properties[CosmosDBDatabaseIDKey] == "") {
+		return nil, fmt.Errorf("missing required properties '%s' and '%s' for an unmanaged resource", CosmosDBAccountIDKey, CosmosDBDatabaseIDKey)
 	}
 
-	// store account so we can delete later
-	properties[CosmosDBAccountIDKey] = *account.ID
-	properties[CosmosDBAccountNameKey] = *account.Name
+	var account *documentdb.DatabaseAccountGetResults
+	var err error
+	if properties[CosmosDBAccountIDKey] == "" {
+		// If we don't have an ID already then we will need to create a new one.
+		account, err = handler.CreateCosmosDBAccount(ctx, properties, documentdb.MongoDB)
+		if err != nil {
+			return nil, err
+		}
 
-	mrc := documentdb.NewMongoDBResourcesClient(handler.arm.SubscriptionID)
-	mrc.Authorizer = handler.arm.Auth
-
-	dbName := properties[CosmosDBNameKey]
-	db, err := handler.CreateDatabase(ctx, *account.Name, dbName, options)
-	if err != nil {
-		return nil, err
+		// store account so we can delete later
+		properties[CosmosDBAccountIDKey] = *account.ID
+		properties[CosmosDBAccountNameKey] = *account.Name
+	} else {
+		// This is mostly called for the side-effect of verifying that the account exists.
+		account, err = handler.GetCosmosDBAccountByID(ctx, properties[CosmosDBAccountIDKey])
+		if err != nil {
+			return nil, err
+		}
 	}
 
-	// store db so we can delete later
-	properties[CosmosDBNameKey] = *db.Name
+	if properties[CosmosDBDatabaseIDKey] == "" {
+		account, err := handler.CreateDatabase(ctx, *account.Name, properties[CosmosDBDatabaseNameKey], options)
+		if err != nil {
+			return nil, err
+		}
+
+		// store db so we can delete later
+		properties[CosmosDBAccountIDKey] = *account.ID
+	} else {
+		// This is mostly called for the side-effect of verifying that the database exists.
+		_, err := handler.GetDatabaseByID(ctx, properties[ServiceBusQueueIDKey])
+		if err != nil {
+			return nil, err
+		}
+	}
 
 	return properties, nil
 }
 
 func (handler *azureCosmosDBMongoHandler) Delete(ctx context.Context, options DeleteOptions) error {
 	properties := options.Existing.Properties
+	if properties[ManagedKey] != "true" {
+		// For an 'unmanaged' resource we don't need to do anything, just forget it.
+		return nil
+	}
+
 	accountName := properties[CosmosDBAccountNameKey]
-	dbName := properties[CosmosDBNameKey]
+	dbName := properties[CosmosDBDatabaseNameKey]
 
 	// Delete CosmosDB Mongo database
 	err := handler.DeleteDatabase(ctx, accountName, dbName)
@@ -73,6 +99,23 @@ func (handler *azureCosmosDBMongoHandler) Delete(ctx context.Context, options De
 	}
 
 	return nil
+}
+
+func (handler *azureCosmosDBMongoHandler) GetDatabaseByID(ctx context.Context, databaseID string) (*documentdb.MongoDBDatabaseGetResults, error) {
+	parsed, err := radresources.Parse(databaseID)
+	if err != nil {
+		return nil, fmt.Errorf("failed to parse CosmosDB Mongo Database resource id: %w", err)
+	}
+
+	mongoClient := documentdb.NewMongoDBResourcesClient(parsed.SubscriptionID)
+	mongoClient.Authorizer = handler.arm.Auth
+
+	account, err := mongoClient.GetMongoDBDatabase(ctx, parsed.ResourceGroup, parsed.Types[0].Name, parsed.Types[1].Name)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get CosmosDB Mongo Database: %w", err)
+	}
+
+	return &account, nil
 }
 
 func (handler *azureCosmosDBMongoHandler) CreateDatabase(ctx context.Context, accountName string, dbName string, options PutOptions) (*documentdb.MongoDBDatabaseGetResults, error) {
